Clean and Organize files for componenets
